requirements.txt
# Core/Base Libraries - Foundational packages that others depend on
pydantic==2.9
python-dotenv

# LangChain Ecosystem - Core framework and extensions
langchain
langchain_community
langchain-text-splitters

# AI/ML Libraries - Machine learning and embedding models
sentence_transformers
langchain-huggingface

# Vector Databases and Storage - Vector storage and retrieval systems
chromadb
faiss-cpu
langchain_chroma

# Document Processing - PDF, web scraping, and content extraction
pypdf
bs4
pymupdf
arxiv
wikipedia

# LLM Providers - Language model integrations
langchain-openai
langchain-ollama
groq
langchain_groq

# Web Frameworks - API and UI frameworks
fastapi
uvicorn
sse_starlette
streamlit
validators

# Development and Runtime Tools
ipykernel
langserve
unstructured
pytube
youtube-transcript-api